ci: generalize recovery gating into a version-agnostic "single-version" suite
Broaden the build tag (recovery -> singleversion) and fold in three more
version-agnostic tests that were needlessly re-running on every MySQL version,
plus the linter suite:

  - TestUniqueOnNonUniqueData  (checksum detecting non-unique data)
  - TestChunkerPrefetching     (chunker large-gap handling)
  - TestUnparsableStatements   (TiDB statement parser)
  - lint_test.go               (linter logic)

These exercise Spirit's own logic, not server datatype behaviour, so they only
need one MySQL version. The three scattered tests move into a new gated
singleversion_test.go (build tags gate per file); lint_test.go and
resume_test.go are tagged in place. The dedicated 8.0.45 job's -run filter is
extended to cover them, and the workflow/service are renamed accordingly.
