+++ pkg/migration/migration_test.go
@@ -130,22 +130,6 @@ func TestRenameInMySQL80(t *testing.T) {
 	require.NoError(t, m.Run())
 }
 
-// TestUniqueOnNonUniqueData tests that we:
-// 1. Fail trying to add a unique index on non-unique data.
-// 2. The error does not blame spirit, but is instead suggestive of user-data error.
-func TestUniqueOnNonUniqueData(t *testing.T) {
-	t.Parallel()
-	tt := testutils.NewTestTable(t, "uniquet1", `CREATE TABLE uniquet1 (id int not null primary key auto_increment, b int not null, pad1 varbinary(1024))`)
-	tt.SeedRows(t, "INSERT INTO uniquet1 (b, pad1) SELECT 1, RANDOM_BYTES(1024)", 100000)
-	testutils.RunSQL(t, `UPDATE uniquet1 SET b = id`)
-	testutils.RunSQL(t, `UPDATE uniquet1 SET b = 12345 ORDER BY RAND() LIMIT 2`)
-
-	m := NewTestMigration(t, WithTable("uniquet1"), WithAlter("ADD UNIQUE (b)"))
-	err := m.Run()
-	require.Error(t, err)
-	require.ErrorContains(t, err, "checksum failed after several attempts. This is likely related to your statement adding a UNIQUE index on non-unique data")
-}
-
 func TestGeneratedColumns(t *testing.T) {
 	t.Parallel()
 	t.Run("unbuffered", func(t *testing.T) {
@@ -293,43 +277,6 @@ func TestStmtWorkflow(t *testing.T) {
 	require.NoError(t, m.Run())
 	// cleanup
 	testutils.RunSQL(t, `DROP TABLE IF EXISTS t1s`)
-}
-
-// TestUnparsableStatements tests that the behavior is expected in cases
-// where we know the TiDB parser does not support the statement. We document
-// that we require the TiDB parser to parse the statement for it to execute,
-// which feels like a reasonable limitation based on its capabilities.
-// Example TiDB bug: https://github.com/pingcap/tidb/issues/54700
-func TestUnparsableStatements(t *testing.T) {
-	t.Parallel()
-	// CREATE TABLE with BLOB DEFAULT — TiDB parser doesn't support this but MySQL does.
-	m := NewTestMigration(t, WithStatement(`CREATE TABLE t1parse (id int not null primary key auto_increment, b BLOB DEFAULT ('abc'))`))
-	require.NoError(t, m.Run())
-
-	// ALTER TABLE with BLOB DEFAULT via --statement — fails because TiDB parser rejects it.
-	m = NewTestMigration(t, WithStatement("ALTER TABLE t1parse ADD COLUMN c BLOB DEFAULT ('abc')"))
-	err := m.Run()
-	require.Error(t, err)
-	require.ErrorContains(t, err, "can't have a default value")
-
-	// ALTER TABLE with BLOB DEFAULT via --table/--alter — works (bypasses parser limitation).
-	m = NewTestMigration(t, WithTable("t1parse"),
-		WithAlter("ADD COLUMN c BLOB DEFAULT ('abc')"))
-	require.NoError(t, m.Run())
-
-	// CREATE TRIGGER — not supported.
-	m = NewTestMigration(t, WithStatement("CREATE TRIGGER ins_sum BEFORE INSERT ON t1parse FOR EACH ROW SET @sum = @sum + NEW.b;"))
-	err = m.Run()
-	require.Error(t, err)
-	require.ErrorContains(t, err, "line 1 column 14 near \"TRIGGER")
-
-	// https://github.com/pingcap/tidb/pull/61498
-	m = NewTestMigration(t, WithTable("t1parse"),
-		WithAlter(`ADD COLUMN src_col timestamp NULL DEFAULT NULL, add column new_col timestamp NULL DEFAULT(src_col)`))
-	require.NoError(t, m.Run())
-
-	// Cleanup
-	testutils.RunSQL(t, `DROP TABLE IF EXISTS t1parse`)
 }
 
 func TestCreateIndexIsRewritten(t *testing.T) {
